Move session initialization to seperate module
