+++ bitcash/network/__init__.py
@@ -5,9 +5,3 @@ from .rates import (
     satoshi_to_currency_cached,
 )
 from .services import NetworkAPI
-
-from requests import Session
-
-#Create a re-usable session object which re-uses connection, mantains a pool.
-session = Session()
-#Optional: Configure user-agent here so that upstream servers know which library/version is making the request
